Fixing squid:S1488, squid:S2147, squid:S2184

+++ api/src/main/java/org/opennms/newts/api/Duration.java
@@ -91,7 +91,7 @@ public class Duration implements Comparable<Duration>, Serializable {
 
             switch (unit) {
                 case "w":
-                    r = Duration.days(value * 7);
+                    r = Duration.days((long)value * 7);
                     break;
                 case "d":
                     r = Duration.days(value);
